+++ libraries/PCA9685_RT/test/unit_test_001.cpp
@@ -76,6 +76,9 @@ unittest(test_constants)
 unittest(test_constructor)
 {
   PCA9685 ledArray(0x20);
+
+  Wire.begin();
+
   assertTrue(ledArray.begin());
   assertTrue(ledArray.isConnected());
   assertEqual(16, ledArray.channelCount());
@@ -85,6 +88,9 @@ unittest(test_constructor)
 unittest(test_TODO)
 {
   PCA9685 ledArray(0x20);
+
+  Wire.begin();
+
   assertTrue(ledArray.begin());
 
   // TODO
@@ -95,12 +101,14 @@ unittest(test_OutputEnable)
 {
   PCA9685 ledArray(0x20);
 
+  Wire.begin();
+
   assertTrue(ledArray.begin());
 
   assertEqual(HIGH, ledArray.getOutputEnable());
 
   assertTrue(ledArray.setOutputEnablePin(12));
-  //  assertEqual(HIGH, ledArray.getOutputEnable()); // need mock
+  //  assertEqual(HIGH, ledArray.getOutputEnable());  //  need mock
 
   assertTrue(ledArray.setOutputEnable(true));
   // assertEqual(LOW, ledArray.getOutputEnable());